Initial port of matrix.py to JavaScript
The goal is to let users generate matrix views on the Attack Flow
website, whereas currently it requires cloning the repo and running some
Python scripts. This first commit ports the existing .py code to .js.

The ported code also gains a new ability: if a technique exists under
multiple tactics, it can use the tactic ID to figure out which cell to
use in the overlay.

In the process, I also updated the Tesla breach to fix one bad technique
ID (it was using mobile technique instead of enterprise) and also add
tactic IDs.