AF-222 get make validate working (#46)
* Build dist version of app

* trying to fix errors

* AF-222 Get the command line script to build and run

Add a vite build target for the CLI (get rid of the CLI tsconfig) and
update the cli.ts script to work with the AF 3.0 libraries. Note that
the corpus is still in 2.0 format so I manually converted Black Basta to
3.0, but I need to convert the rest of the corpus to test the entire
build.

* AF-222 Update cli.ts to work with v3 library

* cli.ts now works with the v3 library
* refactored to use proper argument parsing
* added a subcommand for upgrading v2 files -> v3

* AF-222 Upgrade all existing files in corpus from v2 to v3

I used cli.ts to upgrade all of our existing flows.

* AF-222 Get the build working as much as possible

We have a few more things to fix but all the build steps are roughly
correct at this point and I am getting the build ready to merge to the
public repo.

---------

Co-authored-by: Mark Haase <mhaase@mitre.org>